VIC-8852 Add avs-device-sdk v1.10-anki as dep
Our previous build of `avs-device-sdk` was done locally on a developer's
laptop and pushed up to Artifactory.  We want to be more rigorous and
use our build server to create this package.  A script was added to the
anki/build-tools.git repository to build `avs-device-sdk` and this
change will pull down that result.