BI-1719 Download deps by sha256 hash
I realized that someone may accidentally run one of the dependency
builds again on TeamCity and overwrite the named file, like
pffft-1.0-anki.tar.bz2, with a different archive that has a different
SHA256 value.  This would then break the victor build.  To fix this, I
have made a change to the build-tools' upload-to-s3 script to also copy
the asset to /deps/victor/sha256/<SHA256SUM> so that the file is always
available based on its actual SHA256 hash.  With this change, we will
download from that URL instead of using the name.

For example, pffft-1.0-anki.tar.bz2 is available at:
https://sai-general.s3.amazonaws.com/build-assets/deps/victor/pffft-1.0-anki.tar.bz2

or

https://sai-general.s3.amazonaws.com/build-assets/deps/victor/sha256/681a3b239326719b456af828998af3818f34a1070fa03bc9413acc27bfd7c969