VIC-12977 Fix Habitat relocalize to charger bug (#4839)
Previously the robot would take the distance between the 
identity pose and the current pose as the distance travelled 
since being putdown. This displacement would be used to
test if the robot is not in the habitat (if it drove further than
the diagonal dimension of the tray)

However with relocalizing to the charger, the pose we are
put down at will get rejiggered to be part of an existing
pose tree. As a result we now correctly measure the robot's
displacement from three possible origins:

1. the origin of the frame when we detect being put down
2. the origin of the frame when we get delocalized
3. the origin of the current frame (fallback option)