VIC-13065 Build cloud/go/bin/protoc-gen-*
We were building `protoc-gen-go` and `protoc-gen-grpc-gateway` and
storing them under `~/.anki/protobuf/dist/3.5.1/go/bin`.  This is bad,
because that directory does NOT get cleaned on a clean build.  Rather,
subsequent builds will end up using the same, possibly outdated,
versions of these tools.  Our build of these 2 tools depends on the
version of `go` and the commits checked out from 4 go modules.  With
this change, we will rebuild the tools on every clean build and be up to
date.