VIC-13030 Alexa acoustic test mode (#4847)
A build flag ALEXA_ACOUSTIC_TEST or console var can be used to
put the robot into an acoustic test mode. In acoustic test mode,
once the robot has finished onboarding, if it is on the charger,
it will not move or play any procedural audio or earcons. Only the
"Alexa" wake word is active.