[box] Add console variable system for cloud process (#4884)
To allow the changing of voice services and face configuration
groups for the box demo, this change adds a console variable
system for the cloud process. Any code can register functions
with the `anki/util/cvars` package, and a dedicated endpoint
(`/cvars`) on the cloud web server will allow manipulation of
those variables:

https://i.imgur.com/PmM2ddp.png

As requested, this change also reverts using Microsoft as the
default voice handler.