VIC-11232 DAS event for stack cycle (#4894)
For some das events, the distribution of event counts per robot has
extremely large tails. One of those events is behavior.meet_victor.end,
which would print even when the active feature was not MeetVictor.

I found that if you force a stack cycle by adding a CancelSelf in
OnBehaviorActivated or BehaviorUpdate, robot logs match das logs for
robots in the tail of the distribution, in that behavior.meet_victor.end
is repeatedly printed when another behavior is running.
Another PR is fixing code paths that cause the stack cycle.
This one just adds a das message for stack cycles, once per cycle type
per boot, so that in the future we don't have to waste time figuring out
the cause.
This should fix the large tails in behavior.meet_victor.end and
robot.delocalized, but there are a handful of others left to investigate.