VIC-13070 Fix locale after onboarding (#4978)
Setting the alexa locale would only change the locale if alexa
was running. This caches the locale so it can be applied to alexa
when it is initialized. Also, just in case there's a gap in between
initialization and when the locale gets set, the config files will now
contain the desired locale string.