VIC-5268 [Animation issue?] Eyes show over image displayed on face with DisplayFaceImageRGBChunk CLAD msg (#4852)
Eyes keep alive was still running while the image was displayed, solution
was to insert <duration> ms of disabled keep alive and also set
_wasAnimationInterruptedWithNothing = true. The latter is an override for
the "long enough" timers and the "keep face alive timers" that forces the
face to update asap.

The duration in SetFaceImage() is passed to SpriteSequenceKeyFrame that
correctly re-starts the face alive and some future point. This future
point is immediate with _wasAnimationInterruptedWithNothing set.

Of note, solutions that did not work:
Added a function like ResetKeepFaceAliveTimers() to force the timers
to resume after the face is enabled, resulted in a blank face afte the
image was removed.

Added a function to set _longEnoughSinceLastStreamTimeout_s to force
immediate face alive when face is re-enabled resulted in a blank face
after the image was removed.

Modifying _lastAnimationStreamTime does not work as it is constantly
updated and adding the third/fourth override/timer is not ideal.