VIC-13262: Removed mutex/lock from DropBreadcrumb function (#5003)
* VIC-13262: Vector running Very slow - end listening / animations + audio out of sync

Issues was a mutex lock added in logging code (DropBreadcrumb within logging.cpp).  Since we access this from multiple threads, very often, the threads were blocking like crazy and causing the slowdown.

I removed this mutex, but it should be noted that the data this lock was guarding is still not thread safe.  We should find a thread safe solution for this logging that doesn’t involve blocking.  I don’t know much about this DropBreadcrumb functionality, so I am not in a position to find a solution atm.

I’m banking on the fact that we weren’t crashing prior to this change to determine that it’s “safe enough” to back out this change.