VIC-13199 Give alexa mic input on mac (#4986)
Mic input was not being sent to amazon servers on mac because neither
the alexa microphone component nor the alexa wake word model are given
audio until the latter is initialized. Since there is no mac wake word
engine model, it was never initialized.
This pretends the wake word engine is initialized on mac.