VIC-13259 Dont play alexa error without usage (#5008)
Some robots will suddenly say "Your device has lost its connection"
when first booting, or when it has just been sitting there for some
time.
Most error messages are played if we receive a bad response from the
server when trying to send something. Usually things are only sent
when the user interacts, but it there is some edge case during bootup
where the robot sometimes sends its firmware version before AVS
connects. The fix for that is to not flag an error if the user has
not interacted within the last minute.
A minute is chosen since it can sometimes take ~30 sec from the time
you say the wake word before the SDK responds with an error.
This will _not_ cover the cases where a false positive wake word is
detected and the robot is not connected to the internet.