VIC-13199 Corrected build of avs-device-sdk
Previously, the build of avs-device-sdk for macOS was relying on the
macOS version of libcurl.  This worked fine on Mojave (10.14), but
failed on High Sierra (10.12), because the older macOS lacked support
for http/2.  This new build makes sure to rely on the version of libcurl
from the Homebrew package curl-openssl, which should work on either High
Sierra or Mojave.