VIC-13248 Fix alexa barge in again (#5006)
I'm not sure what changed, but this tunes the barge in detector.
It's now turned off for US and GB since Jordan made the playback
recognizer time indices more accurate. It remains active for AU since
the GB wake word model we use for AU can miss an Aussie's "Alexa."
This also retunes it. The window of frequencies where our speaker can't
output looked smaller than it used to be, so I reduced the range.
To then account for increased variability, that power is now compared
to the power in a range where our speaker can output.

This also updates the speech recog webviz tab
+ Separate ignore columns into reasons why it was ignored
+ Keep a small buffer of recent triggers, similar to the CloudIntents tab