VIC-12392 gpio_create : Export only if needed
As part of tightening up the security around GPIO, we will be making a
change in vicos-oelinux.git to handle exporting GPIOs before services
that need them start up.  In addition, non-root processes will NOT be
allowed to write to `/sys/class/gpio/export`.  This change will first
check if a gpio pin is available before trying to export it.