[VIC-13306] Success/Fail earcon now plays immediately after streaming stops
[VIC-13306] Voice Commands: Move Earcon Immediately After Intent Heard

Previously we waited until after the wakeword listening get-out animation finished before we played our response to the cloud intent success/fail.  This gave us extra time to “listen”, but it is causing a feeling of lagginess.  Now we move the success/fail logic immediately before the get-out is played.