VIC-13354 Check sha256 for downloaded deps
If somebody updates the sha256 checksum in the DEPS file, but does NOT
update the version, we could end up using an older build of a dependency
instead of the newer build.  This is because the deptool was only
checking the contents of the VERSION file in the extracted directory for
the downloaded dependency.  If someone failed to update the version when
they did a new build of a dependency, the deptool would find this old
version and think we were good to go.  With this change, we also store
the sha256 checksum of the downloaded tarball in the extracted directory
and then also require that to match when searching for our dependencies.

Since we have not stored this sha256 checksum in the past, the first
time everyone runs a build with this change, they will have to
re-download dependencies.  That is a slight inefficiency that will save
us long term pain.