VIC-13241 Australian Alexa is deaf after talking about Lisa Wilkinson (#5071)
When using the smaller GB/AU model, certain utterances or Alexa's
response to those utterances leave the recognizer's VAD in a bad state.
After alexa finishes speaking, the vad indicator flickers between ON,
OFF, and back ON again. When this is followed by a quiet wake word, the
VAD immediately switches to OFF, and the wake word is ignored. This is
easiest to reproduce if you switch locales and then play the "Lisa
Wilkinson" clip followed by the "Ireland leader" clip.
The fix disables the VAD for AU/GB.

Also:
+ Special casing for when setting the detection threshold prior to
initializing the detector.
+ Even if PryonLiteDecoder_IsDecoderInitialized(), trying to call
  PryonLiteDecoder_SetDetectionThreshold() without a keyword fails
  with PRYON_LITE_ERROR_NOT_INIT
+ Removes the audio keyframe from the alexa get-out.