VIC-13435 Use sudo /sbin/export-gpio to export gpio pins (#5088)
In this particular bug, it looks like the `mm-qcamera-daemon` service
was crashing and as a result, gpio pin 94 was un-exported.  Eventually,
`vic-engine` noticed this and tried to re-export it by writing `94` to
`/sys/class/gpio/export`. However, `vic-engine` doesn't run as root
anymore.  As a result, when it tries to export a gpio pin by writing to
`/sys/class/gpio/export`, it will fail.  Instead, as of vicos
1.5.0.2912, the `engine` user can use `sudo` to invoke the
`/sbin/export-gpio` script to export pin 94 (camera) and pin 0 (Whiskey
Time of Flight sensor).