VIC-12762 Fix Python SDK stall on custom markers vision mode disable request #5093
Typically, each time a Python SDK script runs, a request is made to turn off custom marker detection (unless the script explicitly wants it on). Sometimes no response is received back from the engine for this request, either in a timely manner or at all. The response we expect is that it was turned off. The current theory is that something else on the robot (not from the SDK) is keeping the custom marker vision mode on.

Given the way the response is currently sent from sdkComponent to gateway, the SDK would only receive the response if custom marker detection is indeed turned off.

We have seen the SDK stall in two ways intermittently after requesting custom marker detection be turned off: either it waits until the response comes back (~50 seconds), or the response never comes back and the SDK hangs indefinitely.

With this change, if custom marker detection is being turned off, we no longer wait for the response from the engine, and instead treat it as a suggestion, and it's OK if it doesn't get turned off. If the SDK script is turning it on, we wait for confirmation that it was turned on.

Also reenable the CustomMarker test in `test_all_messages.py`.