VIC-13369 Fix mutex/lock from DropBreadcrumb function (#5055)
* VIC-13369 Fix mutex/lock from DropBreadcrumb function

breadcrumbs are not thread-safe, adding mutexes introduces too much
contention that results in poor performance; most notably with
animations.

This uses a single static buffer that is shared between all threads
for a given process. But, uses thread-local storage to reference
into this buffer. Each thread has it's own area in this buffer
and writes that same information as previously in a round-robin
fashion.

The storage is pre-allocated. If the number of threads exceeds the
buffer victor will assert in debug. In release the thread will
use a previously allocated amount of buffer, there will be
conflict between the two threads but it won't crash.

* Allow thread space to wrap, tests exceed 34 threads.