VIC-13454 dependencies.py: Defer URL checking
If we don't need to download something, don't check if the URL is up.
If the URL is only available on the local LAN or over VPN, this check
can cause the build to hang when on a remote network.

In addition, I found that if we failed to contact an SVN server or
artifact server, we would simply print a warning instead of raising an
error and quitting.  This could allow an invalid build to proceed and
waste developer time.  I converted this warnings to RuntimeErrors so
that the build is halted immediately.

In addition, to cut down on the time it takes to determine that we
cannot contact our internal Artifactory server, I changed the timeout
for the `is_up` routine to be 10 seconds.