VIC-7903 Add animation name to PlayAnimationResponse #5110
PlayAnimationResponse is returned for both PlayAnimationRequest and PlayAnimationTriggerRequest. Added partly for parity with Cozmo.

Originally for this ticket I planned to implement Python events like Cozmo had for animation (e.g., 'EvtAnimationsLoaded', 'EvtAnimationTriggersLoaded'), and in researching that I found this missing animation name in the response that Cozmo had. Shawn showed that the events are not needed, so adding this anim name is all that results from the investigation.

One minor cool thing that comes of it is that if you play a trigger, the specific anim that was selected comes back in this response.

This does not cause a backwards compatible issue for Python SDK 0.5.1.