VIC-13504 Handle case of uncached blobstore key
The previous code would mistakenly try to download the blobstore key
again if it was under `EXTERNALS`, but not in the cache location.  As a
result, even if you had the key, we would try to download it again.  The
URL requires being on the local LAN or having VPN access and thus your
build would fail.

To fix this, we now check both locations for the file.  In addition, the
DEPS file now contains a sha256 checksum for the file to make sure that
what we downloaded or cached is what we expect.

To note, the previous code was chopping off the newline terminator for
the file.  The new code does not.  Therefore, the first time you run
this new code, even if you have the old 22 byte key file, we will
re-download the correct 23 byte key file.  If you really don't want to
download 23 bytes, you can append a newline to your old
`EXTERNALS/victor-blobstore-qalogs-key`.