VIC-13455 Verify tflite files at build time (#5096)
Sometimes developers get into a wonky state and these tflite files that
are downloaded with git lfs do not get properly downloaded.  Don't know
the true root cause of this, but this is an added check for our tflite
files.  If they don't have the magic string `TFL3` in them, then they
are probably bogus.  If we detect that, we stop the build and give
suggestions on how to fix.