VIC-13382: Add support for measuring and graphing vic-robot latency
This patch adds support for measuring and graphing the vic-robot loop duration
and jitter using LTTNG.

This patch adds the convert.py script which requires the babeltrace. The
convert.py script reads in an LTTNG trace and writes out a trace file that can
be directly read into python. This allows us to run the convert script inside a
docker image then graph using the native graphing libraries on your host
machine.

The graph_durations.py and graph_latency.py scripts are written to read in a
trace file created by the convert.py script.

In order to visualize this data run the convert.py script on the trace directory
to create a trace file. Once the trace file is created you can use the
graph_latency.py or graph_durations.py script to visualize the measurements.