VIC-13478 Play error clip after playlist 404 (#5129)
When amazon provides a valid URL of a playlist with an unavailable
source, the UrlContentToAttachmentConverter spins up a thread
to call onError, which we never implemented.
This change sets a flag to play the "I'm having trouble
playing this song" clip. It's not perfect, because we don't know
what sourceID had the error, and we don't know when onError gets
called relative to play/pause/stop/etc. This should hopefully cover
most cases though.