VIC-9888 Add /anki/etc/victor-compat-version
We don't want to deploy victor.git code onto a robot with an
incompatible OS.  In the past, we simply checked the OS version (minus
the build number) to see if they matched and if they didn't we blocked
the deploy.  However, that required updating the OS version for every
incompatible change.  They caused trouble for the release engineering
and product management teams who were planning around what the OS
version was going to be for an upcoming release.

With this change and the corresponding change in vicos-oelinux.git, we
have a new internal-use only compatibility version that we check when
doing deploys so that we don't deploy incompatible code.  When a change
is made to vicos-oelinux that breaks compatibility, the
`VICTOR_COMPAT_VERSION` in both repos will need to be updated.