ToF range status check
Adds "background" ToF range status check. Once enabled, ranging will be cycled on and off periodically. The range data is monitored for any ROIs that are reporting consistant range errors. If an ROI is found to be consistantly reporting errors, the test is considered to have failed and a callback is called.
The goal of this test is to catch any electrical issues with the decoupling capacitor or flex cable.

The test makes sure to not interfere with any external requests for ranging to be enabled such as playpen. Example, if someone has requested ranging and the test has entered the stop_ranging cycle, then it will not actually stop ranging.