VIC-13715: Add helpers to use Docker on OSX
This patch adds a helper script to build a victor docker image for developers.
The run.sh script creates an image named victor-${USER} extended from the victor
image specified by docker/Dockerfile. The victor-${USER} image is created on
demand with a USER:UID matching the current user.

The home directory of the current user is mounted as the home directory in the
container to make it easy to share files.

The Dockerfile.dev is where developers should add additional packages needed for
development.