VIC-13436 Tell Chewie if we are an anki.dev bot (#5195)
If we are running a shipping build on one of our internal `anki.dev`
robots, we want Chewie to give us an OTA URL for a dev-signed version of
the shipping build.  However, if we just return an unmodified
`ro.build.id`, Chewie doesn't know if we are a production robot for end
users or if we are an `anki.dev` robot.  With this change, we append
`-ankidev` to the end of the build id string so that Chewie will know.
This change will need to be back ported for Whiskey robots' recovery
partition.

This is a back port of commit 87197bfc89 from `master`.