VIC-13734 Fix wiggle back when leaving charger (#5205)
The battery level can drop from Full to Nominal the moment that the
robot leaves the charger contacts. When the robot starts to drive
off the charger, StayOnChargerUntilCharged can then want to activate,
causing the robot to wiggle back again.
The fix ensures that if the battery drops from Full to Nominal, the
behavior only activates after enough time is spent at Nominal for
the robot to clear the charger platform.