VIC-13830 Get tensorflow from outside of VPN
Previously, we had to be on the local LAN or VPN to get the tensorflow
artifact from an internal Artifactory server.  In addition, this
artifact was slightly out of date as it was built with a prior version
of the vicos sdk.  With this change, we get the tensorflow artifact,
built with the current vicos-sdk, whether we are on the local LAN or
not.