VIC-13835 Make it harder to deploy to incompatible robot
Require individual flags to ignore compatibility mismatch and version
mismatch.  Use red blinking text to express to user what a bad idea it
is.  Hopefully, this will reduce these deploys to only cases where a
developer absolutely knows that the incompatibility is not an issue for
what they are testing.