VIC-13125 Wrapper namespace support for clad emitters (#4977)
In some scenarios, it may be convenient to add a global "wrapper"
namespace to all generated code. This is useful in cases where different
emitters for the same lang are used on an identical set of CLAD files,
or if CLAD files are included in libraries that are intended for distribution
and linking.  In these cases use of an emitter-defined namespace avoid symbol
resolution conflicts that occur during linking in C languages.

This change set adds a special emitter that mutates the AST after the parsing
step in order to add a namespace, if specified. This strategy enables the namespace
to be added to any emitter that supports namespaces.

The initial use case for this is to add a namespace for generated CPPLite code
in order to avoid name/symbol conflicts if generated CPP code is also present.