VIC-13742 Self test cube connection failure (#5212)
Cube connections are unreliable so ignore the results of the connection check in the self test. The wifi check is still enabled so that will cover the radio check for the warehouse scenario (will actually need to make a change for the special self-destructing warehouse build as currently the wifi check result is masked by the ignore cube connection flag).

Currently unsure of what to do for the Whiskey self test in factory-0.9.1. If the cube connections are somehow better there then fine this check can be enabled there. Otherwise I have to check the radio in someway since the same self test code will be run by both users and the warehouse. Theoretically could gate the check off the existence of a file or something but ideal the warehouse would not need to pair with the robot.

* Switching out hardcoded self test ssid and password with randomly generated ones