VIC-13882 Retry cube connection before timeout
After we initiate a connection to the cube, we will sometimes get an
unexpected disconnect callback from `ankibluetoothd`.  This is due to a
low level BLE connection problem between Vector and his cube.  Not
always, but sometimes, if we try to connect again, we will succeed.
This code adds in that retry logic.  Ultimately, once we reach our
timeout, we will give up and fail.