Removed the minimum stream duration for a normal (non-faked) stream
** note, there still is the concept of “fake streaming” … it is used by the wakeword training behavior, as well as when the battery is low

This should be changed into more of a state machine, but for now this will eliminate the min streaming for a “normal” wakeword.  It will stop “streaming” as soon as we get word from the cloud that we’re done, or a timeout (as usual) and allow us to stream something immediately aftewards.