VIC-13860 Disambiguate clad structures using CppLite namespace
CLAD generated CPP lite structures are created in a CppLite namespace,
this change modifies the hand-authored code to reference those
structures in that namespace. However, because the code is shared
with non-CPP lite generated CLAD structures all accesses must
be wrapped in a macro (CLAD and variants) that is used depending
which library of which CLAD structures is used.

Adds a USES_CPPLITE #define to all CppLite projects, this a public
property the propagates up the hierarchy. If USES_CPPLITE is defined
but generated/clad/engine is in the include path then there is a
mixture of CppLite and non-CppLite in the target.

Re-organizes and splits robot and engine structures for coretech,
e.g. generated/clad/robot/coretech/common, generated/clad/robot/coretech/vision,
generated/clad/engine/coretech/common, generated/clad/engine/coretech/vision

Re-organizes faceDetectionMetaData.clad to meet byte alignment requirements,
this needs checking for compatibility problems.