VIC-14082 Fix frequent charger deletion in GoHome Lowlight
The charger was being deleted because it was being considered as "old" because its observation times were not being updated while it was being localized to. As a result, the charger once seen would not have up to date age, leading to deletion by the behavior GoHome (especially during lowlight), despite being recognized and detected.