VIC-13454 dependencies.py: Defer URL checking (#5282)
Updating svn_checkout() with support to easily disable the RuntimeError if needed when the SVN server is unavailable (because there are times when developers don't necessarily need the latest assets from SVN but they want their build to succeed anyway). To take advantage of that, developers can locally set 'abort_if_offline' to False when building without access to the SVN server.