VIC-13989 Fix Robot Getting Stuck In Keepaway (#5349)
It looks like BlockWorld may be cycling objects in and out of its
Localized list a lot quicker than before. If the cube of interest in
keepaway is connected, then there's the potential to get into a corner
case where the target is invalid (unknown by blockworld) but still
recently seen. This particular combo was causing an infinitely looping
state cycle in keepaway and intermittent observations of the target in
the mix were preventing keepaway timeouts from triggering.

This fix ignores target validity for a moment provided the target has
been recently seen, but this merits some additional effort some time
in the future to ensure keepaway is properly accounting for
BlockWorld's intricacies.