VIC-3300: Localize BlackJack behavior & strings (#5358)
1. Add localizable string file for blackjack strings.

2. Revise BehaviorBlackJack to use (1)

I put blackjack strings into a separate localization file because they may be shared with other behaviors.

The blackjack behavior depends on some additional user prompt behaviors that will be localized in a separate PR.

Some of the translations are trivial, e.g. score is just converted to numeric string, but it seemed prudent to use a localization key in case we need to do further customization for non-English translations.