VIC-13126 Python SDK: Expose MeetVictor behavior (#5392)
MeetVictor is required for enroll a face, which we are exposing in the SDK.

Also move AssumeBehaviorControl down as Python SDK does not use it. Not sure why it was exposed in SDK; presumably that was done in error. An public SDK user has used it, which we don't want.

In the future need to make adding a behavior less boilerplate copy/paste code.