VIC-11720 Always use an Agent for connman
For the factory-0.9.1 branch of vicos-oelinux, we have upgraded
connman from 1.30 to 1.36.  Some things changed with this upgrade.
Namely, we need to explicitly enable WiFi before we start scanning
for access points and we need to use the Agent interface when
configuring a WiFi access point.  This change does some minor
refactoring of our WiFi code and back ports the changes from the
`master` branch to always use the Agent interface.