VIC-14265 Fix delay on silence intents (#5416)
The PromptUserForVoiceCommand behavior was lingering in the listening
state upon receiving the `silence` intent. These changes handle that
specific intent directly by transitioning immediately into a reprompt
if `silence` is received. This makes sense for the prompt behavior
since the logic that gets skipped compared to an Unknown intent
doesn't really apply if the user simply hasn't said anything.