VIC-13521 Allow engine to remap SpriteBoxes at runtime (#5451)
Add the ability to start running a canned animation with a list of
SpriteBoxKeyFrame asset remaps in a single message. This paves the way
to restructure Weather, Timer, Clock, and BlackJack around
SpriteBoxKeyFrame infrastructure, alleviating dependencies on the
CompositeImage system, and allowing for use of animation groups with
these features.