VIC-14365: Button not responding well in calm mode (#5459)
This bug was introduced when #5140 was merged which removes syscon calm mode. It used to be the case that when syscon was in calm mode it would kick itself out of calm mode temporarily whenever the backpack button was pressed which would make RobotState messages send more frequently such that double-press could be detected. This PR implements a similar mechanism to temporarily disable calm mode (which now just refers to the state of sending RobotState messages slowly) for 1 second whenever the button is pressed.

Also removed some unused head and lift functions