VIC-14346 Disable autoexposure not working (#5449)
Due to the default schedule for autoexposure and whitebalance in visionScheduleMediator_config.json the modes would always run even if "disabled" by VisionComponent.

Fix is to remove the default schedule and instead have VisionComponent request those modes be enabled on init. This allows disabling to work by simply unsubscribing from the two vision modes.

Added special handling for AutoExposure and WhiteBalance modes in visionScheduleMediator to prevent them from being balanced in the schedule. This allows them to run at the same time on the same image and prevents them from fighting with each other when applying camera settings.