VIC-14367 docs/DEPS.md - Explain '3rd' directory
We don't want to use our own proprietary `deptool` and `DEPS` file if we
can avoid it.  This documentation update explains how to build and
integrate a 3rd party dependency under our `3rd` directory.  Most
importantly it covers the use of `git LFS` to avoid storing large binary
objects in the git repository.