WIP combined CLAD and CPPLite emitter
A CPP_CPPLite_emitter.py; all of the things from CLAD and all of
the things from CPPLite with the members from CPPLite (I also
experimented with the members from CLAD). Looks like CLAD but
with the memory layout of CPPLite, IsValid(), Size() from CPPLite,
enumToString functions from both.

Members from CLAD isn't useful, as it's the one feature we need
from CPPLite, the memory layout.

Members from CPPLite is useful, but that makes some of the methods
from CLAD not useful e.g. the constructors. Constructors don't work,
because we're passing strings to the constructors and the initialization
doesn't work (from std::vector to char [])

It's not finished, tags don't work and many other things don't work,
it's going to take time to convert, which might just end up being
time converting anim to CLAD. So I'm going to pause and try one of two things:
1) switch anim to CLAD and see how much work it is
2) write a clang plugin to automatically re-write our code