SAIP-2107: Command line tool for remote robot fleet management
This is needed when we want to remotely control a canary robot
fleet. In that case this tool will be used by the build server
for instance to redeploy the fleet when new robot code is merged.

This can be done via the admin host nd ssh or by exposing the
commands through some RPC mechanism (e.g. HTTP / gRPC).