VIC-7566 Add a shared mem vic-anim -> gateway audio pipeline.
Add a shared memory circular buffer that allows vic-anim to pipe
Signal-Essence-processed data to the gateway via shared memory.
There is a client-server implementation in C++, as well as a
golang client.

For the moment, there is no locking on the reader count. This
should be addressed if this system is going to be used for
future projects (with multiple readers.)