VIC-13522 Refactor Weather and BlackJack for SpriteBoxCompositor (#5481)
BlackJack refactor temporarily points BlackJack at a set of Hand
edited animations used to add SpriteBox keyframes to the original
BlackJack animations so that they could be remapped at run time to
display the pertinent assets.  This alleviates all of BlackJacks
dependencies on the CompositeImage system, and BlackJack no longer
requires any of its animations to be loaded on the engine process.

* Refactor weather
These changes add a batch of hand reworked weather animations which
add the necessary SpriteBoxKeyFrames for weather to be handled by the
SpriteBoxCompositor.  This way ALL the pertinent information for a
weather behavior is centralized to the appropriate animation instead
of its orignal state of being distributed across a bunch of JSON files
and requiring anims to be loaded to the engine process for timestamp
extraction.