VIC-14463 Uncalibrated lift is bouncy (#5488)
Cherry-pick of f4e1bea32c3ca2fcf92d3f9240574f5b5ba8a0e3 from cozmo-one

An uncalibrated lift would oscillate repeatedly if moved due to the removal of the D term from the PID controller when the lift was near its limits. The removal of the D term was originally added to prevent bouncing of the lift when the lift was moved rapidly down, COZMO-10221. In order to preserve the fix for that issue but also address the bouncing when uncalibrated, I have changed the logic for when to remove the D term. Instead of removing it whenever the current angle is outside the max/min lift angle, it is now only removed within a range of the max/min lift angle (5 degrees above the max or 5 degrees below the min).