VIC-7566 Switch stream back to big endian.
Yesterday, we switched to little endian in an attempt to fix up issues
was hearing in his SDK test. The reason that worked is because the
default for that library was to treat audio data in little endian
order. Since there's no compelling reason to be either little or
big... go big or go home.