VIC-14028 VIC-14404 VIC-14444 Defensive OTA status messages
We can get into some tricky states with OTA updates and progress
reporting:
1. The previous update failed
2. The user wants to start a new update, but we are in the middle
   of an existing one
3. The hourly update is sleeping (aka delaying) before it gets
   started, but the user wants to start right now
4. The current 'progress' or 'expected' values are temporarily invalid
   during a change in state, like in a delta OTA when transitioning
   from downloading to applying.

This change adds defensive code to handle each of those scenarios.  In
addition, a slight tweak was made to the update-engine to set the
initial progress to 1 instead of 0 when handling a delta file.  This
tweak helps get the Chewie UI to transition sooner.