VIC-14513 Centralize SpriteBox Definitions and use AssetIDs (#5500)
These changes pull the definitions of SpriteBoxes and
SpriteBoxKeyframes down to the CLAD level so that they can be easily
constructed and sent from the engine to the active animation. To
simplify that pipeline, we also switch from using string assetNames at
the endpoints to using hashed string assetID's throughout both
processes, converted from strings at load time. Definitions of
AssetID's generally, and specific special-case assetID's are also
centralized under the SpritePathMap to simplify changes should they
become necessary.

These changes pave the way to convert Timer/Clock to use the
SpriteBoxCompositor infrastructure, which will alleviate the last
dependencies on the CompositeImage system. Once that is done,
CompositeImages can be reduced to basic collections of the same
SpriteBoxes mentioned above with appropriate rendering logic, and a
significant amount of code and complexity can be removed from the
AnimationStreamer and the CompositeImage classes themselves.