VIC-13522 Clock and Timer refactor (#5509)
These changes refactor the Clock and Timer behaviors to leverage the
SpriteBoxCompositor framework. Instead of sending CompositeImages to
update the clock, SpriteBoxKeyFrames are constructed and batch sent to
the AnimationStreamer for insertion into the running animation. This
significantly reduces the message data load for the features, and
removes the last engine-side dependencies on the CompositeImage
system, which can now be refactored for increased simplicity and
narrowed scope of responsibility.