VIC-14427: Petting doesn't register until you pet him on table top first (#5485)
This was because the touch sensor wasn't calibrating when picked up. This was to avoid calibrating falsely to a likely touched state, but since we have a fast calibration mechanism that kicks in if the baseline drops it seems ok to take out the picked up restriction. He now calibrates anytime, but you still may need to not pet him for few seconds after picking up before petting can be detected.